\name{Downscale-package}
\alias{downscale-package}
\docType{package}
\title{
Downscaling species occupancy
}
\description{
A set of functions that downscale species occupancy from coarse grain sizes to 
predict species occupancy at finer grain sizes.
}
\details{
\tabular{ll}{
Package: \tab downscale\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-27\cr
License: \tab GPL-2\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Charles Marsh

Maintainer: Charles Marsh <charliem2003@gmail.com>
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
