% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Downscale.R
\name{downscale}
\alias{downscale}
\title{Model area of occupancy against grain size for downscaling}
\usage{
downscale(occupancy, area, model, extent)
}
\arguments{
\item{AOO}{Observed area of occupancy}

\item{Area}{Grain size}

\item{Model}{Downscaling model (see Details)}
}
\description{
Model area of occupancy against grain size for downscaling
}

