% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.downscale.R, R/predict.downscale.R
\name{predict.downscale}
\alias{plot}
\alias{plot.downscale}
\alias{predict}
\alias{predict.downscale}
\title{Predict occupancy at fine grain sizes}
\usage{
\method{plot}{predict.downscale}(predict.object, ...)

predict(object, newdata, extent, tolerance = 1e-6, plot = FALSE)
}
\arguments{
\item{predict.object}{an object from \code{\link{predict.downscale}}.}

\item{object}{a fitted object of class \code{'downscale'}.}

\item{newdata}{vector of grain sizes (in squared units e.g. km^{2}) for which
area of occupancy will be predicted.}

\item{extent}{total area in same units as newdata (only required for
\code{FNB} and \code{Thomas} models).}

\item{tolerance}{only applicable for the \code{Thomas} model. The tolerance
used during integration in the Thomas model during optimisation of
parameters. Lower numbers allow for greater accuracy but require longer
processing times (default = \code{1e-6}).}

\item{plot}{if \code{plot = TRUE} (default = \code{FALSE}) plots observed and
  predicted occupancies against grain size on a log-log plot.}

\item{\dots}{arguments, including graphical parameters for plot.downscale,
passed to other methods.}
}
\value{
Returns an object of class \code{'predict.downscale'} with three
  objects:
   \item{model}{Downscaling model used.}
   \item{predicted}{Data frame containing two columns:
   \tabular{lll}{
     \code{Cell.area} \tab  \tab Grain sizes for which occupancy have been
       estimated\cr
     \code{Occupancy} \tab  \tab Predicted area of occupancy for each grain
       size\cr
     }
   }
   \item{observed}{Data frame containing two columns:
   \tabular{lll}{
     \code{Cell.area} \tab  \tab Grain sizes for which occupancy have been
       observed\cr
     \code{Occupancy} \tab  \tab Observed area of occupancy for each grain
       size\cr
     }
   }
}
\description{
Predict area of occupancy at fine grain sizes using parameters
  in a downscale object estimated from coarse grain sizes using
  \code{\link{downscale}}. There is also a simple \code{\link{plot}}
  function.
}
\details{
The function takes the parameters for a downscaling model estimated
  through \code{\link{downscale}} and uses the model to predict area of
  occupancy at finer grain sizes. See \code{\link{downscale}} for details on
  the downscaling models and their parameterisation. Plotting can be called
  directly from \code{\link{predict.downscale}} or from
  \code{\link{plot.downscale}}.
}
\examples{
some examples
}
\author{
Charles Marsh <\email{charliem2003@gmail.com}> with input from
  Louise Barwell.
}
\references{
Azaele, S., Cornell, S.J., & Kunin, W.E. (2012). Downscaling
  species occupancy from coarse spatial scales. \emph{Ecological
  Applications} 22, 1004-1014.

Barwell, L.J., Azaele, S., Kunin, W.E., & Isaac, N.J.B. (2014).
  Can coarse-grain patterns in insect atlas data predict local occupancy?
  \emph{Diversity and Distributions} 20, 895-907.
}
\seealso{
See \code{\link{downscale}} for estimating parameters of a
  downscaling function from observed occupancies at coarse grain sizes
}

